requirements.txt
langchain==1.2.15
langchain-community==0.4.1
langchain-text-splitters==1.1.1
openai==2.30.0
faiss-cpu==1.13.2
numpy==2.2.6
python-dotenv==1.0.0
streamlit==1.32.0
